+++ prisma.config.ts
@@ -8,4 +8,7 @@ export default defineConfig({
   datasource: {
     url: process.env.DATABASE_URL,
   },
+  migrations: {
+    seed: 'ts-node prisma/seed.ts',
+  },
 });